# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type NewsCategory = typeof NewsCategory[keyof typeof NewsCategory];


export const NewsCategory = {
  international: 'international',
  india: 'india',
  'united-states': 'united-states',
  technology: 'technology',
} as const;

export type NewsPerspective = typeof NewsPerspective[keyof typeof NewsPerspective];


export const NewsPerspective = {
  left: 'left',
  center: 'center',
  right: 'right',
  unknown: 'unknown',
} as const;

export type CoverageSegmentPerspective = typeof CoverageSegmentPerspective[keyof typeof CoverageSegmentPerspective];


export const CoverageSegmentPerspective = {
  left: 'left',
  center: 'center',
  right: 'right',
} as const;

export interface CoverageSegment {
  perspective: CoverageSegmentPerspective;
  /** @minimum 0 */
  count: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  percentage: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  width: number;
}

export interface CoverageSummary {
  /** @minimum 0 */
  total: number;
  segments: CoverageSegment[];
}

export interface NewsArticle {
  id: string;
  title: string;
  description: string;
  url: string;
  source: string;
  sourceUrl: string;
  /** @nullable */
  publishedAt: string | null;
  category: NewsCategory;
  imageUrl?: string;
  perspective: NewsPerspective;
  coverage?: CoverageSummary;
}

export interface NewsListResponse {
  articles: NewsArticle[];
  /** @minimum 0 */
  failedSourceCount: number;
}

export interface ApiError {
  error: string;
}

export type GetNewsParams = {
category?: NewsCategory;
};