# lib/api-zod/src/generated/types/coverageSegment.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CoverageSegmentPerspective } from './coverageSegmentPerspective';

export interface CoverageSegment {
  perspective: CoverageSegmentPerspective;
  /** @minimum 0 */
  count: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  percentage: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  width: number;
}