# lib/api-zod/src/generated/types/coverageSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CoverageSegment } from './coverageSegment';

export interface CoverageSummary {
  /** @minimum 0 */
  total: number;
  segments: CoverageSegment[];
}