# lib/api-zod/src/generated/types/newsArticle.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CoverageSummary } from './coverageSummary';
import type { NewsCategory } from './newsCategory';
import type { NewsPerspective } from './newsPerspective';

export interface NewsArticle {
  id: string;
  title: string;
  description: string;
  url: string;
  source: string;
  sourceUrl: string;
  /** @nullable */
  publishedAt: Date | null;
  category: NewsCategory;
  imageUrl?: string;
  perspective: NewsPerspective;
  coverage?: CoverageSummary;
}