# lib/api-zod/src/generated/types/newsCategory.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type NewsCategory = typeof NewsCategory[keyof typeof NewsCategory];


export const NewsCategory = {
  international: 'international',
  india: 'india',
  'united-states': 'united-states',
  technology: 'technology',
} as const;