# lib/api-zod/src/generated/types/newsListResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NewsArticle } from './newsArticle';

export interface NewsListResponse {
  articles: NewsArticle[];
  /** @minimum 0 */
  failedSourceCount: number;
}