# lib/api-zod/src/generated/types/newsPerspective.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type NewsPerspective = typeof NewsPerspective[keyof typeof NewsPerspective];


export const NewsPerspective = {
  left: 'left',
  center: 'center',
  right: 'right',
  unknown: 'unknown',
} as const;